refactor(core): move sensitive packages from pkg to internal
